Move configuration class into separate file
